backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
sqlmodel==0.0.8
pymysql==1.1.0
python-jose==3.3.0
cryptography==41.0.4
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
requests==2.31.0
markdown==3.5.1
sseclient-py==1.7.2 